requirements.txt
bidict==0.22.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
netifaces==0.10.6
python-engineio==4.4.0
python-socketio==5.8.0
PyYAML==6.0
six==1.16.0
Werkzeug==2.2.3
